requirements.txt
streamlit
numpy
pandas
plotly
selenium
beautifulsoup4
requests
stanza
Pillow
matplotlib
fitz
google-generativeai